test: update intro_cog tests
